+++ selection/randomized/tests/test_selective_MLE_onedim.py
@@ -5,7 +5,7 @@ from scipy.stats import norm as ndist
 import matplotlib.pyplot as plt
 import nose.tools as nt
 
-from ..lasso import highdim
+from ..lasso import lasso
 from ...tests.instance import gaussian_instance
 from statsmodels.distributions import ECDF
 
@@ -17,11 +17,11 @@ def test_onedim_lasso(n=50000, W=1.5, signal=2., sigma=1, randomizer_scale=1):
         X /= np.sqrt((X**2).sum(0))[None, :]
         Y = X.dot(beta) + sigma * np.random.standard_normal(n)
 
-        conv = highdim.gaussian(X,
-                                Y,
-                                W * np.ones(X.shape[1]),
-                                randomizer_scale=randomizer_scale * sigma,
-                                ridge_term=0.)
+        conv = lasso.gaussian(X,
+                              Y,
+                              W * np.ones(X.shape[1]),
+                              randomizer_scale=randomizer_scale * sigma,
+                              ridge_term=0.)
 
         signs = conv.fit()
         nonzero = signs != 0
